+++ laws/src/main/scala/spire/laws/shadows/ShadowAdditiveCSemigroup.scala
@@ -1,0 +1,8 @@
+package spire.laws.shadows
+
+import spire.algebra.AdditiveCSemigroup
+
+trait ShadowAdditiveCSemigroup[A, S] extends AdditiveCSemigroup[Shadow[A, S]] with ShadowAdditiveSemigroup[A, S] {
+  implicit def A: AdditiveCSemigroup[A]
+  implicit def S: AdditiveCSemigroup[S]
+}
