+++ laws/src/main/scala/spire/laws/shadows/ShadowMultiplicativeCSemigroup.scala
@@ -1,0 +1,8 @@
+package spire.laws.shadows
+
+import spire.algebra.MultiplicativeCSemigroup
+
+trait ShadowMultiplicativeCSemigroup[A, S] extends MultiplicativeCSemigroup[Shadow[A, S]] with ShadowMultiplicativeSemigroup[A, S] {
+  implicit def A: MultiplicativeCSemigroup[A]
+  implicit def S: MultiplicativeCSemigroup[S]
+}
